feat: added the migration to app/ folder, need to solve all the errors and the missing css
