+++ app/drive/programs/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { VStack } from '@chakra-ui/react';
 
 import DriveCards from 'components/cards/DriveCards';
@@ -6,18 +8,17 @@ import Navigation from 'components/navigation/Navigation';
 
 import { useDriveContext } from 'contexts/drive';
 
-const Shared = (): JSX.Element => {
-	const { sharedFiles } = useDriveContext();
-	const { sharedPrograms } = useDriveContext();
+const Programs = (): JSX.Element => {
+	const { programs } = useDriveContext();
 
 	return (
 		<Navigation>
 			<VStack w="100%" spacing="48px" align="start">
-				<LabelBadge label="Share with me" />
-				<DriveCards files={sharedFiles.filter((elem) => !elem.deletedAt)} programs={sharedPrograms} />
+				<LabelBadge label="My programs" />
+				<DriveCards programs={programs} />
 			</VStack>
 		</Navigation>
 	);
 };
 
-export default Shared;
+export default Programs;
