+++ package.json
@@ -15,13 +15,14 @@
     "axios": "^1.4.0",
     "boring-avatars": "^1.10.1",
     "crypto-js": "^4.1.1",
+    "encoding": "^0.1.13",
     "eth-crypto": "^2.6.0",
     "ethers": "^5.7.2",
     "framer-motion": "^10.12.17",
     "git-clone": "^0.2.0",
     "js-file-download": "^0.4.12",
-    "next": "^12.3.4",
-    "next-auth": "4.10.3",
+    "next": "^13.4.11",
+    "next-auth": "^4.22.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
@@ -63,6 +64,7 @@
     "eslint": "^8.43.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-next": "^13.4.10",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-jsx-a11y": "^6.7.1",
