+++ src/components/computing/github/GithubModal.tsx
@@ -1,7 +1,7 @@
 import { HStack, Input, Skeleton, Text, useColorModeValue, useToast, VStack } from '@chakra-ui/react';
 import axios from 'axios';
 import { signIn, signOut, useSession } from 'next-auth/react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { ChangeEvent, useEffect, useState } from 'react';
 
 import Button from 'components/Button';
