+++ src/components/navigation/Navigation.tsx
@@ -1,5 +1,5 @@
 import { VStack } from '@chakra-ui/react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 
 import { useConfigContext } from 'contexts/config';
