--- src/components/navigation/layoutResponsiveBar.tsx
@@ -1,13 +1,0 @@
-import {ResponsiveBar} from "./ResponsiveBar";
-
-const ResponsiveBarLayout = ({children}: { children: React.ReactNode, }) => {
-
-  return (
-    <>
-      <ResponsiveBar/>
-      {children}
-    </>
-  )
-}
-
-export default ResponsiveBar;
