+++ tsconfig.json
@@ -20,11 +20,18 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "preserve",
-    "incremental": true
+    "incremental": true,
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
   "include": [
     "src",
-    "pages"
+    "app",
+    "pages",
+    ".next/types/**/*.ts"
   ],
   "exclude": [
     "node_modules"
